src/api/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref>=3.8.1,<4
billiard==4.2.0
celery==5.4.0
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
Django==5.1.14
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-dump-die==0.1.5
django-recaptcha==4.0.0
django-rest-knox==5.0.1
django-stubs-ext==5.0.2
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
kombu==5.3.7
mypy-extensions==1.0.0
packaging==24.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.32.4
setuptools==78.1.1
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.6.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.42.0
